Fix bug for TimeoutManager timeouts 24+ days in the future
Because setTimeout can only accept a 32-bit signed int in node, trying to schedule
a timeout that's too far out will result in it being scheduled for right now.
Instead, handle this case by saving the timeout and scheduling it either on a future
reboot or in an emergency fallback timeout in 10 days (in case the runtime lives that long).