Optimize A-Star pathfinding
Since the heuristics we use are consistent (i.e. a successor node will never result in a lower total cost),
we can assume that the first time we visit a location will always be the cheapest path to that location.
Thus, we never have to visit a given location twice.