Add FileRecordStorage utility
This utility writes rows of JSON records to a JSONL file, utilizing the OS's ability to append to files
without having to open and parse the entire file every time. This will be useful in cases where large
quantities of records need to be frequently written, yet rarely read. The utility will support in-place
purging to prevent file bloat.